pose_classification/requirement.txt
albumentations==0.4.5
astroid==2.11.7
autopep8==1.7.0
axial-positional-embedding==0.2.1
black==22.6.0
blessed==1.19.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
configparser==5.2.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
dill==0.3.4
easydict==1.9
einops==0.4.1
filterpy==1.4.5
flake8==5.0.4
Flask==2.0.3
flatbuffers==2.0.7
gpustat==1.0.0
graphviz==0.8.4
h5py==2.10.0
hickle==4.0.4
idna==3.3
imageio==2.15.0
imgaug==0.2.6
importlib-metadata==4.2.0
importlib-resources==5.4.0
isort==4.3.21
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
json-logging==1.3.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
local-attention==1.4.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mxnet-cu101==1.9.1
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.19.5
nvidia-ml-py==11.495.46
onnxruntime==1.10.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
product-key-memory==0.1.10
protobuf==3.19.4
psutil==5.9.2
pycodestyle==2.9.1
pyflakes==2.5.0
pylint==2.13.9
pyparsing==3.0.9
python-dateutil==2.8.2
pytorchcv==0.0.67
PyWavelets==1.1.1
PyYAML==6.0
reformer-pytorch==1.4.4
requests==2.27.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
tifffile==2020.9.3
toml==0.10.2
tomli==1.2.3
tqdm==4.64.0
typed-ast==1.4.3
typing_extensions==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.11.2
XlsxWriter==3.0.3
zipp==3.6.0
